# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * The Padel Exchange API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegistrationInputIndustry = typeof RegistrationInputIndustry[keyof typeof RegistrationInputIndustry];


export const RegistrationInputIndustry = {
  Technology: 'Technology',
  Financial_Services: 'Financial Services',
  Professional_Services: 'Professional Services',
  'Cyber_/_Security': 'Cyber / Security',
  Legal: 'Legal',
  Consulting: 'Consulting',
  Healthcare: 'Healthcare',
  Other: 'Other',
} as const;

export type RegistrationInputFunction = typeof RegistrationInputFunction[keyof typeof RegistrationInputFunction];


export const RegistrationInputFunction = {
  'Founder_/_CEO': 'Founder / CEO',
  'Risk_/_Compliance_/_GRC': 'Risk / Compliance / GRC',
  'Security_/_CISO': 'Security / CISO',
  'Product_/_Engineering': 'Product / Engineering',
  'Sales_/_Marketing': 'Sales / Marketing',
  Operations: 'Operations',
  Investor: 'Investor',
  Other: 'Other',
} as const;

export type RegistrationInputSeniority = typeof RegistrationInputSeniority[keyof typeof RegistrationInputSeniority];


export const RegistrationInputSeniority = {
  'Founder_/_Owner': 'Founder / Owner',
  'C-suite': 'C-suite',
  'VP_/_Head_of': 'VP / Head of',
  'Director_/_Manager': 'Director / Manager',
  Other: 'Other',
} as const;

export type RegistrationInputPadelLevel = typeof RegistrationInputPadelLevel[keyof typeof RegistrationInputPadelLevel];


export const RegistrationInputPadelLevel = {
  Never_played: 'Never played',
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
} as const;

export type RegistrationInputInterestsItem = typeof RegistrationInputInterestsItem[keyof typeof RegistrationInputInterestsItem];


export const RegistrationInputInterestsItem = {
  'Playing_/_fitness': 'Playing / fitness',
  Meeting_other_founders: 'Meeting other founders',
  'Industry_peers_&_ideas': 'Industry peers & ideas',
  Just_trying_padel: 'Just trying padel',
} as const;

export interface RegistrationInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  company?: string;
  jobTitle?: string;
  industry?: RegistrationInputIndustry;
  function?: RegistrationInputFunction;
  seniority?: RegistrationInputSeniority;
  padelLevel?: RegistrationInputPadelLevel;
  interests?: RegistrationInputInterestsItem[];
  linkedinUrl?: string;
  gdprConsent: boolean;
}

export interface Registration {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  function?: string | null;
  /** @nullable */
  seniority?: string | null;
  /** @nullable */
  padelLevel?: string | null;
  /** @nullable */
  interests?: string[] | null;
  /** @nullable */
  linkedinUrl?: string | null;
  gdprConsent: boolean;
  createdAt: string;
}

export type ListRegistrationsParams = {
adminPassword: string;
};

export type ExportRegistrationsParams = {
adminPassword: string;
};