# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * The Padel Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: number;
  fullName: string;
  email: string;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  function?: string | null;
  /** @nullable */
  seniority?: string | null;
  /** @nullable */
  padelLevel?: string | null;
  /** @nullable */
  interests?: string[] | null;
  /** @nullable */
  linkedinUrl?: string | null;
  gdprConsent: boolean;
  createdAt: Date;
}