# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * The Padel Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputFunction } from './registrationInputFunction';
import type { RegistrationInputIndustry } from './registrationInputIndustry';
import type { RegistrationInputInterestsItem } from './registrationInputInterestsItem';
import type { RegistrationInputPadelLevel } from './registrationInputPadelLevel';
import type { RegistrationInputSeniority } from './registrationInputSeniority';

export interface RegistrationInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  company?: string;
  jobTitle?: string;
  industry?: RegistrationInputIndustry;
  function?: RegistrationInputFunction;
  seniority?: RegistrationInputSeniority;
  padelLevel?: RegistrationInputPadelLevel;
  interests?: RegistrationInputInterestsItem[];
  linkedinUrl?: string;
  gdprConsent: boolean;
}